# docs/openapi.yaml
openapi: 3.0.3
info:
  title: LoanApp API
  version: 1.0.0
  description: Frozen v1 API contract (all registered routes listed).
servers:
  - url: /api/v1
...
paths:
...
  /me/bank-accounts:
    get:
      tags: [BankAccounts]
      summary: List bank accounts
      security: [{ bearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Ok' }
    post:
      tags: [BankAccounts]
      summary: Create or update bank account
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GenericRequest' }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
...
  /admin/borrowers:
    get:
      tags: [AdminBorrowers]
      summary: List borrowers
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Query' }
      responses:
        '200': { $ref: '#/components/responses/PaginatedOk' }
    post:
      tags: [AdminBorrowers]
      summary: Create borrower
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GenericRequest' }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
  /admin/borrowers/{id}:
    get:
      tags: [AdminBorrowers]
      summary: Borrower details
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/IdPath' }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
  /admin/borrowers/{id}/risk:
    get:
      tags: [AdminBorrowers]
      summary: Borrower risk
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/IdPath' }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
  /admin/borrowers/{id}/notes:
    post:
      tags: [AdminBorrowers]
      summary: Add borrower note
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/IdPath' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GenericRequest' }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
  /admin/borrowers/{id}/override:
    put:
      tags: [AdminBorrowers]
      summary: Set borrower override
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/IdPath' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GenericRequest' }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
  /admin/roles:
    get:
      tags: [AdminRoles]
      summary: List roles
      security: [{ bearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Ok' }
  /admin/admin-users/{id}/roles:
    put:
      tags: [AdminRoles]
      summary: Assign role
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/IdPath' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GenericRequest' }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  headers:
    X-Request-Id:
      description: Request correlation ID
      schema: { type: string }
  parameters:
...
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, default: 50, minimum: 1, maximum: 200 }
    Cursor:
      name: cursor
      in: query
      required: false
      schema: { type: string }
    Query:
      name: query
      in: query
      required: false
      schema: { type: string }
...
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  schemas:
...
    PaginatedEnvelope:
      type: object
      properties:
        items:
          type: array
          items: { type: object, additionalProperties: true }
        nextCursor: { type: string, nullable: true }
      required: [items, nextCursor]
    GenericRequest:
      type: object
      additionalProperties: true
    GenericObject:
      type: object
      additionalProperties: true
...
  responses:
...
    Ok:
      description: Generic OK response
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/GenericObject' }
    PaginatedOk:
      description: Generic paginated response
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/PaginatedEnvelope' }